Corrigindo ProfessorService

+++ src/main/java/com/project/professor/allocation/service/ProfessorService.java
@@ -1,60 +1,71 @@
 package com.project.professor.allocation.service;
 
 import java.util.List;
+
 import org.springframework.stereotype.Service;
+
+import com.project.professor.allocation.entity.Department;
 import com.project.professor.allocation.entity.Professor;
 import com.project.professor.allocation.repository.ProfessorRepository;
 
 @Service
 public class ProfessorService {
 
-	private ProfessorRepository professorRepository;
+	private final ProfessorRepository professorRepository;
+	private final DepartmentService departmentService;
 
-	public ProfessorService(ProfessorRepository professorRepository) {
+	public ProfessorService(ProfessorRepository professorRepository, DepartmentService departmentService) {
 		super();
 		this.professorRepository = professorRepository;
+		this.departmentService = departmentService;
 	}
 
-	// CRUD: Read All
-	public List<Professor> findAll() {
-		List<Professor> professor = professorRepository.findAll();
-		return professor;
+	public List<Professor> findAll(String name) {
+		if (name == null) {
+			return professorRepository.findAll();
+		} else {
+			return professorRepository.findByNameContainingIgnoreCase(name);
+		}
 	}
 
-	// CRUD: Read by ID
 	public Professor findById(Long id) {
 		return professorRepository.findById(id).orElse(null);
 	}
 
-	// CRUD: Create
-	public Professor create(Professor professor) {
-		professor.setId(null);
-		professor = professorRepository.save(professor);
-		return professor;
+	public List<Professor> findByDepartment(Long departmentId) {
+		return professorRepository.findByDepartmentId(departmentId);
 	}
 
-	// CRUD: Update
+	public Professor save(Professor professor) {
+		professor.setId(null);
+		return saveInternal(professor);
+	}
+
 	public Professor update(Professor professor) {
 		Long id = professor.getId();
 		if (id != null && professorRepository.existsById(id)) {
-			professor = professorRepository.save(professor);
+			return saveInternal(professor);
 		} else {
 			return null;
 		}
-		return professor;
 	}
 
-	// CRUD: Delete By Id
 	public void deleteById(Long id) {
-		if (professorRepository.existsById(id)) {
+		if (id != null && professorRepository.existsById(id)) {
 			professorRepository.deleteById(id);
 		}
-
 	}
 
-	// CRUD: Delete All
 	public void deleteAll() {
+		professorRepository.deleteAllInBatch();
+	}
 
-		professorRepository.deleteAll();
+	private Professor saveInternal(Professor professor) {
+		professor = professorRepository.save(professor);
+
+		Department department = departmentService.findById(professor.getDepartmentId());
+		professor.setDepartment(department);
+
+		return professor;
 	}
 }
